‘111 ’

+++ unitree_rl_lab/deploy/include/isaaclab/assets/articulation/articulation.h
@@ -36,7 +36,7 @@ struct ArticulationData
 
     Eigen::Quaternionf root_quat_w;
 
-    std::vector<float> joint_ids_map;
+    std::vector<int> joint_ids_map;
 
     unitree::common::UnitreeJoystick* joystick = nullptr;
 };
